% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPTeCD.R
\name{OPTeCD}
\alias{OPTeCD}
\title{Optimal Partial Tetra-Allele Cross Designs for Prime Number of Lines}
\usage{
OPTeCD(v)
}
\arguments{
\item{v}{Any prime number (>3)}
}
\value{
This function generates a series of universally optimal family of designs using MOLS. The method starts with selecting any of the (N-1)/2  MOLS of a given order N (the number of lines) and retaining the first four rows and making crosses with the lines occurring in each column. The parameters of the developed class of design is total number of crosses (T) =N*(N-1)/2, number of blocks (b) = (N-1)/2, number of replications (r), block sizes (k) = N and  degree of fractionation (f)=4/(N-2)(N-3)
}
\description{
Optimal Partial Tetra-Allele Cross Designs for Prime Number of Lines
}
\examples{
library(OPTeCD)
OPTeCD(5)
}
